requirements.txt
django==3.2.5
django-phonenumber-field==7.1.0
django-simpleui==2023.3.1
djangorestframework==3.14.0
django-filter==23.2
Pillow==9.5.0
PyMySQL==1.1.0
pyyaml==6.0
uritemplate==4.1.1
coreapi==2.3.3
drf-yasg==1.21.6